Adds the base for every strategy used to classify/counter log lines
This is the base (suggestion) for every counter used to inspect log
lines parsed by the gem. The main ideia is to make easier to change,
add or remove strategies used to inspect the log lines.

For instance: You're using a Counter which only count paths which
contains some query strings. So the concret method 'count(line)'
will contain the logic to filter lines. Also, it can use a different
regex to filter lines as well.