Adds the LinePresenter
This class is responsible to display the result of the Counter.
It's another design decision which makes possible to add, remove, and
change the way that we display the result of the Parser + Counter.

The Presenter itself must accept the rows provided by the Counter
and a description.

The LinePresenter simply creates a string containing the description
plus every row (path + visits).